// src/app/models/sjcs/DestinatariosRetencItem.ts
export class DestinatariosRetencItem {
  idinstitucion: string;
  nombre: string;
  iddestinatario: string;
  cuentacontable: string;
  historico: boolean;
  fechabaja: Date;
  orden: string;
  nombrepartidatemp: string;
  descripciontemp: string;
  cuentacontabletemp: string;
  constructor() { }
}